Increased max_digits on unit_amount_decimal to 24 from 19
Updated max_digits from 19 to 24 for unit_amount_decimal and amount_decimal model fields.
This was done because the largest allowed value by Stripe is for the IDR (Indonesian Rupiah)
which is 999,999,999,998.999,999,999,999 (24 digits, 12 decimals).
In fact the Stripe console allows one to enter up to 14 decimals but
automatically rounds off to 12 decimal places in its API.